% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centend-package.R
\docType{package}
\name{centend-package}
\alias{centend}
\alias{centend-package}
\title{centend: Examine Property Ownership Over Time}
\description{
This library groups property owners by shared owner names and addresses, and business officers. It aims to create a total portfolio of properties for these grouped entities.
}
\author{
\strong{Maintainer}: Renz Torres \email{wr3ns@proton.me}

}
\keyword{internal}
