% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_owners_to_corps.R
\name{connect_owners_to_corps}
\alias{connect_owners_to_corps}
\title{Match owner names to corporate filings}
\usage{
connect_owners_to_corps(owner_names, filings, id, name)
}
\arguments{
\item{owner_names}{a vector of owner names}

\item{filings}{a data frame consisting of corporate filings}

\item{id}{a unique identifying column within filings}

\item{name}{the name column in filings}
}
\value{
a crosswalk data frame of owner_names to corporate filings, may be one-to-many
}
\description{
Matches a vector of owner names to corporate filings
}
\details{
Using a table of corporate filings, finds corporate filings that
match strings with owner names
}
