% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_properties.R
\name{create_property_time_blocks}
\alias{create_property_time_blocks}
\title{Create property nodes by time}
\usage{
create_property_time_blocks(sales, parcels)
}
\description{
This method creates a true representation of property owned
by a single entity over time using sales data. For example, it will divide
a parcel P sold in 2008, 2012, and 2019 into four properties, 2000 (or
whatever the beginning of the target period is)-2008,
2008-2012, 2012-2019, and 2019-2020 (or whatever the end target period is).
}
