% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{florida_places}
\alias{florida_places}
\title{A list of Florida places}
\format{
\subsection{\code{florida_places}}{

A data frame with 964 rows and 3 columns:
\describe{
\item{County}{The county in which the place is found}
\item{Place}{The name of the place}
\item{Type}{City, town, or unincorporated census-designated place}
}
}
}
\source{
United States Census Bureau, 2020 geographies
}
\usage{
florida_places
}
\description{
Places as defined by the United States census refers to incorporated cities,
towns, and villages, as well as unincorporated but important named localities.
This list was compiled from the 2020 United States census list of geographies.
}
\keyword{datasets}
