% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_nooh.R
\name{identify_nooh}
\alias{identify_nooh}
\title{Identify not-owner-occupied status}
\usage{
identify_nooh(
  data,
  owner_address,
  owner_city,
  owner_state,
  owner_zip,
  physical_address,
  physical_zip,
  homestead,
  n_units,
  records,
  .state_strings = c("FL", "FLORIDA", "FLA"),
  .city_strings
)
}
\arguments{
\item{data}{input data frame}
}
\value{
a modified input data frame with two vectors
}
\description{
Identifies single-family rentals and second homes. Outputs two
vectors, one a status of "Owner-occupied", "Not owner-occupied", and "Unknown",
and another as to why something was categorized accordingly.
}
