% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_parcels.R
\name{identify_parcels}
\alias{identify_parcels}
\title{Uniquely identify parcels and parcel owners}
\usage{
identify_parcels(parcels, parcel_id, parcel_year, owner_name, owner_address)
}
\arguments{
\item{parcels}{a data frame}

\item{parcel_id}{the column containing the parcel id}

\item{parcel_year}{the column containing the parcel year}

\item{owner_name}{the column containing the owner name}

\item{owner_address}{the column containing the owner address. Should be a single column
that corresponds to the irreducible owner address search string, e.g. a street
address and city already combined
based on hashing methods}
}
\value{
A data frame with unique identifiers
}
\description{
Takes a data frame of parcels and owners and grants unique identifiers
}
\details{
this function is designed on a one-to-one mapping from
a parcel and a single owner. It should still provide similar results on
one-to-many mappings but it is yet untested.
}
