% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_agents.R
\name{match_agents}
\alias{match_agents}
\title{Match agents in addresses}
\usage{
match_agents(parcels, agents, address, city)
}
\arguments{
\item{parcels}{a tagged output from \code{identify_parcels}}

\item{agents}{a table of registered agent addresses}

\item{address}{the address column in agents}

\item{city}{the city column in agents}
}
\value{
parcels with a modified agent column
}
\description{
Takes output from \code{identify_parcels} and matches against
a table of registered agent addresses. If an owner address matches
a registered agent address, return true
}
\details{
Every column provided in ... must provide true when
}
