% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_corp_officers.R
\name{match_corp_officers}
\alias{match_corp_officers}
\title{Find officers of corporations}
\usage{
match_corp_officers(corp_ids, officers, join_id)
}
\arguments{
\item{corp_ids}{a vector of corporate ids,}

\item{officers}{a table of officers}

\item{join_id}{a column in officers corresponding to corp_ids}
}
\value{
a data frame with officers
}
\description{
matches a vector of corporation ids to officers
}
\details{
This function assumes a normalized business registry where
officers are located in another table than the filings
}
