% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cyphers.R, R/cypher_templates.R
\name{read_cypher}
\alias{read_cypher}
\title{Create Cypher queries}
\usage{
read_cypher(file)

read_cypher(file)
}
\description{
Using glue templates, create queries for working with Neo4j.

Create Cypher import and export scripts using glue scripts. Each script has
specific parameters to fill in.
}
