% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_or.R
\name{regex_or}
\alias{regex_or}
\title{Regex 'or'}
\usage{
regex_or(..., as_word = TRUE)
}
\arguments{
\item{...}{Strings for the or statement. Can be strings or regex expressions.}

\item{as_word}{Whether to return a regex or that's detectable only as an isolated word.}
}
\value{
A regex expression
}
\description{
Create a regex 'or' statement, with the option of bookending it as a word.
}
\examples{
# example code
regex_or("Alachua", "Gainesville", "GAINESVILLE")
regex_or(chopper("Gainesville", "Gainesv"), chopper("Alachua County", "Alachua C"))
}
