% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_attn.R
\name{remove_attn}
\alias{remove_attn}
\title{Nix care-ofs and attentions}
\usage{
remove_attn(x)
}
\arguments{
\item{x}{An address vector to be cleaned}
}
\value{
a vector where each entry containing C/O or ATTN has been
transformed into an NA
}
\description{
In datasets where address is split between two fields,
nullify the field that contains a C/O or ATTN
}
\details{
This function is based off a Shimberg-specific application where
address is split in between two fields but one field maximum is required
by the analysis. C/Os and ATTNs make it harder to pattern match with
a registered agents field so this function takes care of it.

This is an optional function in the centend version of this script.
}
