% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimbergize.R
\name{shimbergize}
\alias{shimbergize}
\title{Utility renaming functions}
\usage{
shimbergize(data, what, from)
}
\arguments{
\item{data}{Parcel or sales data.}

\item{what}{One of \code{"parcels"} or \code{"sales"}.}

\item{from}{One of \code{"fdor"} or \code{"ec2"}.}
}
\value{
The cleaned data source
}
\description{
This is a function that applies Shimberg Center business rules to
column names of data from either the Florida Department of Revenue or
Shimberg's preprocessing server. Not really meant for use in the real world.
}
