% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_owners.R
\name{tag_owners}
\alias{tag_owners}
\title{Tag owners}
\usage{
tag_owners(owner_name)
}
\arguments{
\item{owner_name}{A character vector of owner names}
}
\value{
A character vector with types
}
\description{
Categorize owners as one of Corporate, Trust, Governmental,
Trust services, Confidential, or Other depending on aspects of the name.
}
\examples{
# example code
categorize_owner(c("fyr residential llc", "anne ray trust", "564 Newell Tr llc"))
}
