% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transfer_codes}
\alias{transfer_codes}
\title{Real property transfer codes}
\format{
\subsection{\code{transfer_codes}}{

A data frame with 34 rows and 4 columns:
\describe{
\item{code}{Code found in Florida Sales Data File}
\item{short_desc}{A short description of the sale type}
\item{long_desc}{The verbatim description of the sale type from the Florida Department of Revenue}
\item{distressed}{Whether a sale type is one that is generally made under duress}
}
}
}
\source{
https://floridarevenue.com/property/Documents/salequalcodes_bef01012019.pdf
}
\usage{
transfer_codes
}
\description{
Codes used by the Florida Department of Revenue to describe qualities of sales
as, for example, arm's length, bank transfers, foreclosures, under duress, etc.
}
\keyword{datasets}
