% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_components.R
\name{visualize_components}
\alias{visualize_components}
\title{Visualize owner entity graphs}
\usage{
visualize_components(id, records, output = "./temp.html", open = TRUE, ...)
}
\arguments{
\item{id}{A single id. Multiple ids are supported but results don't look good.}

\item{records}{A records object from a wcc query.}

\item{output}{Name of output html file. Defaults to \code{./temp.html}}

\item{open}{Whether to automatically open in a browser. Defaults to TRUE.}

\item{...}{Other parameters to pass to \link[networkD3]{simpleNetwork}.}
}
\value{
The string to the output file, invisibly.
}
\description{
Pass a vector of component ids and visualize connections between
names and addresses. Useful for debugging connection errors.
}
