% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopper.r
\name{chopper}
\alias{chopper}
\title{Detectable regex substring expressions}
\usage{
chopper(full_string, min_string)
}
\arguments{
\item{full_string}{The entire string}

\item{min_string}{The minimum string from which to be expanded}
}
\value{
a regex expression
}
\description{
Creates a regex expression that detects a minimum truncated substring,
and additional letters up to the full string.
Useful for matching string columns where names might be truncated.
}
\examples{
# example code
chopper("Jacksonville", "Jacksonv")
chopper("identifiable", "identifi")
}
