% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_parcels.R
\name{identify_parcels}
\alias{identify_parcels}
\title{Uniquely identify parcels and parcel owners}
\usage{
identify_parcels(
  parcels,
  parcel_id,
  parcel_year,
  owner_name,
  owner_address,
  owner_city
)
}
\arguments{
\item{parcels}{A data frame}

\item{parcel_id}{The column containing the parcel id}

\item{parcel_year}{The column containing the parcel year}

\item{owner_name}{The column containing the owner name}

\item{owner_address}{The column containing the owner street address}

\item{owner_city}{The column containing the owner city}

\item{...}{Any other columns to include, can be renamed through this interface}
}
\value{
A data frame with unique identifiers
}
\description{
Takes a data frame of parcels and owners and grants unique identifiers
}
\details{
this function is designed on a one-to-one mapping from
a parcel and a single owner. It should still provide similar results on
one-to-many mappings but it is yet untested.
}
