% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{load_records}
\alias{load_records}
\alias{tagged_properties}
\alias{component_summary}
\title{Postprocessing functions}
\usage{
load_records(path)

tagged_properties(records)

component_summary(records)
}
\description{
These are functions to operate on a json export result of a
Weakly-Connected-Components result from Neo4j. Really just some shortcuts
}
\section{Functions}{
\itemize{
\item \code{load_records()}: Load records

\item \code{tagged_properties()}: Tag properties

\item \code{component_summary()}: Summarize results

}}
